requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
pydantic
python-dotenv
Faker 